wrote some database schemas and a function for saving an event that also produces a map for the event's master image using helper functions in mapmaker.js.
